rename func module to function
