+++ components/dada-ir/src/code/validated.rs
@@ -4,7 +4,7 @@
 
 use crate::{
     class::Class,
-    func::Function,
+    function::Function,
     in_ir_db::InIrDb,
     intrinsic::Intrinsic,
     op::Op,
